+++ CHANGELOG.md
@@ -4,4 +4,6 @@ Syndi Changelog
 unreleased
 ----------
 
++ csyndi/logger.c: `Syndi::Logger#debug(message)` added. NOTE: This is purely for consistency
+  with Celluloid. Use `#verbose` instead.
 + csyndi/logger.c: `Syndi::Logger#deprecate(message)` added.